선수 프로필 정보 크롤링, 등번호 없는경우에 null값으로(데이터 유실오류 처리 완료)

+++ DB.sql
@@ -27,7 +27,7 @@ CREATE TABLE player (
 	teamName	CHAR(100) NOT NULL,
 	height	INT(5)	NOT NULL,
 	weight	INT(5)	NOT NULL,
-	`number`	INT(5)	NULL,
+	`number`	INTEGER(5)	NULL,
 	`position`	CHAR(10)	NOT NULL,
 	birthDate	CHAR(20)	NOT NULL,
 	profileImage	VARCHAR(255)	NULL	COMMENT 'NULL 허용 사진 없는 경우도 있기 때문에',
